Did more comprehensive refactoring
