+++ src/realtime_api_async_python/audio/bidirectional_audio.py
@@ -1,0 +1,84 @@
+import pyaudio
+import queue
+import asyncio
+from ..utils.logging import logging
+
+# Audio recording parameters
+CHUNK = 1024
+FORMAT = pyaudio.paInt16
+CHANNELS = 1
+RATE = 24000
+
+class BidirectionalAudio:
+    def __init__(self):
+        self.p = pyaudio.PyAudio()
+        self.stream = self.p.open(
+            format=FORMAT,
+            channels=CHANNELS,
+            rate=RATE,
+            input=True,
+            frames_per_buffer=CHUNK,
+            stream_callback=self.callback,
+        )
+        self.queue = queue.Queue()
+        self.is_recording = False
+        self.is_receiving = False
+        logging.info("AsyncMicrophone initialized")
+
+    def callback(self, in_data, frame_count, time_info, status):
+        if self.is_recording and not self.is_receiving:
+            self.queue.put(in_data)
+        # if self.is_recording:
+        #     self.queue.put(in_data)
+        return (None, pyaudio.paContinue)
+
+    def start_recording(self):
+        self.is_recording = True
+        logging.info("Started recording")
+
+    def stop_recording(self):
+        self.is_recording = False
+        logging.info("Stopped recording")
+
+    def start_receiving(self):
+        self.is_receiving = True
+        self.is_recording = False
+        logging.info("Started receiving assistant response")
+
+    def stop_receiving(self):
+        self.is_receiving = False
+        logging.info("Stopped receiving assistant response")
+
+    def get_audio_data(self):
+        data = b""
+        while not self.queue.empty():
+            data += self.queue.get()
+        return data if data else None
+
+    def close(self):
+        self.stream.stop_stream()
+        self.stream.close()
+        self.p.terminate()
+        logging.info("AsyncMicrophone closed")
+
+async def play_audio(audio_data):
+    p = pyaudio.PyAudio()
+    stream = p.open(format=FORMAT, channels=CHANNELS, rate=RATE, output=True)
+    stream.write(audio_data)
+
+    # Add a small delay (e.g., 100ms) of silence at the end to prevent popping, and weird cuts off sounds
+    silence_duration = 0.2  # 200ms
+    silence_frames = int(RATE * silence_duration)
+    silence = b"\x00" * (
+        silence_frames * CHANNELS * 2
+    )  # 2 bytes per sample for 16-bit audio
+    stream.write(silence)
+
+    # Add a small pause before closing the stream to make sure the audio is fully played
+    await asyncio.sleep(0.5)
+
+    stream.stop_stream()
+    stream.close()
+    p.terminate()
+    logging.debug("Audio playback completed")
+
