+++ src/realtime_api_async_python/openai/chat_prompt.py
@@ -1,0 +1,26 @@
+import openai
+import os
+
+def chat_prompt(prompt: str, model: str) -> str:
+    """
+    Run a chat model based on the specified model name.
+
+    Args:
+        prompt (str): The prompt to send to the OpenAI API.
+        model (str): The model ID to use for the API call.
+
+    Returns:
+        str: The assistant's response.
+    """
+    client = openai.OpenAI(api_key=os.getenv("OPENAI_API_KEY"))
+
+    completion = client.beta.chat.completions.parse(
+        model=model,
+        messages=[
+            {"role": "user", "content": prompt},
+        ],
+    )
+
+    message = completion.choices[0].message
+
+    return message.content
