+++ src/realtime_api_async_python/openai/structured_output_prompt.py
@@ -1,0 +1,35 @@
+
+import os
+import openai
+from pydantic import BaseModel
+
+
+def structured_output_prompt(prompt: str, response_format: BaseModel) -> BaseModel:
+    """
+    Parse the response from the OpenAI API using structured output.
+
+    Args:
+        prompt (str): The prompt to send to the OpenAI API.
+        response_format (BaseModel): The Pydantic model representing the expected response format.
+
+    Returns:
+        BaseModel: The parsed response from the OpenAI API.
+    """
+    client = openai.OpenAI(api_key=os.getenv("OPENAI_API_KEY"))
+
+    completion = client.beta.chat.completions.parse(
+        model="gpt-4o-2024-08-06",
+        messages=[
+            {"role": "user", "content": prompt},
+        ],
+        response_format=response_format,
+    )
+
+    message = completion.choices[0].message
+
+    if not message.parsed:
+        raise ValueError(message.refusal)
+
+    return message.parsed
+
+
